+++ src/routes/views/details/Details.scss
