+++ src/routes/views/details/Details.tsx
