+++ src/routes/views/overview/components/CommittedSpend/CommittedSpend.scss
