+++ src/routes/views/overview/components/CommittedSpend/CommittedSpend.tsx
@@ -1,4 +1,4 @@
-import './ActualSpendBreakdown.scss';
+import './CommittedSpend.scss';
 
 import { getQuery, parseQuery, Query } from 'api/queries/query';
 import { Report } from 'api/reports/report';
@@ -10,17 +10,17 @@ import { connect } from 'react-redux';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 import { EmptyValueState } from 'routes/components/state';
 import { NotAvailable } from 'routes/state';
-import { ReportSummary } from 'routes/views/Overview/components/ReportSummary';
+import { ReportSummary } from 'routes/views/overview/components/ReportSummary';
 import { createMapStateToProps, FetchStatus } from 'store/common';
 import { dashboardSelectors, DashboardWidget } from 'store/dashboard';
 import { reportActions, reportSelectors } from 'store/reports';
 import { formatCurrency } from 'utils/format';
 
-interface ActualSpendBreakdownOwnProps {
+interface CommittedSpendOwnProps {
   widgetId: number;
 }
 
-interface ActualSpendBreakdownStateProps {
+interface CommittedSpendStateProps {
   query: Query;
   queryString: string;
   report?: Report;
@@ -29,17 +29,17 @@ interface ActualSpendBreakdownStateProps {
   widget: DashboardWidget;
 }
 
-interface ActualSpendBreakdownDispatchProps {
+interface CommittedSpendDispatchProps {
   fetchReport: typeof reportActions.fetchReport;
 }
 
-export type ActualSpendBreakdownProps = ActualSpendBreakdownStateProps &
-  ActualSpendBreakdownOwnProps &
-  ActualSpendBreakdownDispatchProps &
+export type CommittedSpendProps = CommittedSpendStateProps &
+  CommittedSpendOwnProps &
+  CommittedSpendDispatchProps &
   RouteComponentProps<void> &
   WrappedComponentProps;
 
-const ActualSpendBreakdownBase: React.FC<ActualSpendBreakdownProps> = ({
+const CommittedSpendBase: React.FC<CommittedSpendProps> = ({
   fetchReport,
   intl,
   queryString,
@@ -52,13 +52,19 @@ const ActualSpendBreakdownBase: React.FC<ActualSpendBreakdownProps> = ({
     fetchReport(widget.reportPathsType, widget.reportType, queryString);
   }, [queryString]);
 
-  let value: string | React.ReactNode = <EmptyValueState />;
+  let balance: string | React.ReactNode = <EmptyValueState />;
+  let committed: string | React.ReactNode = <EmptyValueState />;
 
   const hasTotal = report && report.meta && report.meta.total;
   const hasCost = hasTotal && report.meta.total.cost && report.meta.total.cost.total;
 
   if (hasTotal) {
-    value = formatCurrency(
+    balance = formatCurrency(
+      hasCost ? report.meta.total.cost.total.value : 0,
+      hasCost ? report.meta.total.cost.total.units : 'USD',
+      {}
+    );
+    committed = formatCurrency(
       hasCost ? report.meta.total.cost.total.value : 0,
       hasCost ? report.meta.total.cost.total.units : 'USD',
       {}
@@ -73,10 +79,10 @@ const ActualSpendBreakdownBase: React.FC<ActualSpendBreakdownProps> = ({
         <NotAvailable />
       ) : (
         <>
-          <div>August 2022 - February 2023</div>
+          <div>March 2023 - July 31, 2023</div>
           <div className="valueContainer">
-            <div className={`value`}>{value}</div>
-            <div>{intl.formatMessage(messages.overLastMonth)}</div>
+            <div className={`value`}>{balance}</div>
+            <div>{intl.formatMessage(messages.outOf, { value: committed })}</div>
           </div>
         </>
       )}
@@ -85,7 +91,7 @@ const ActualSpendBreakdownBase: React.FC<ActualSpendBreakdownProps> = ({
 };
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-const mapStateToProps = createMapStateToProps<ActualSpendBreakdownOwnProps, ActualSpendBreakdownStateProps>(
+const mapStateToProps = createMapStateToProps<CommittedSpendOwnProps, CommittedSpendStateProps>(
   (state, { widgetId }) => {
     const widget = dashboardSelectors.selectWidget(state, widgetId);
     // const queries = dashboardSelectors.selectWidgetQueries(state, widgetId);
@@ -123,9 +129,9 @@ const mapStateToProps = createMapStateToProps<ActualSpendBreakdownOwnProps, Actu
   }
 );
 
-const mapDispatchToProps: ActualSpendBreakdownDispatchProps = {
+const mapDispatchToProps: CommittedSpendDispatchProps = {
   fetchReport: reportActions.fetchReport,
 };
 
-const ActualSpendBreakdown = withRouter(ActualSpendBreakdownBase);
-export default injectIntl(connect(mapStateToProps, mapDispatchToProps)(ActualSpendBreakdown));
+const CommittedSpend = withRouter(CommittedSpendBase);
+export default injectIntl(connect(mapStateToProps, mapDispatchToProps)(CommittedSpend));
