+++ src/routes/views/overview/components/CommittedSpendTrend/CommittedSpendTrend.scss
